Doc corrections
Execution tests of test.py
Built-in product now a const

+++ test.py
@@ -3,9 +3,11 @@
 """
 import argparse
 import os
+import sys
 import time
+from typing import NoReturn, Optional
 
-from dm_api.dm_api import DmApi, DmApiRv
+from dm_api.dm_api import DmApi, DmApiRv, TEST_PRODUCT_ID
 
 # Get configuration from the environment.
 # All the expected variables must be defined...
@@ -15,6 +17,19 @@ KEYCLOAK_REALM: str = os.environ['SQUONK_API_KEYCLOAK_REALM']
 KEYCLOAK_CLIENT_ID: str = os.environ['SQUONK_API_KEYCLOAK_CLIENT_ID']
 KEYCLOAK_USER: str = os.environ['SQUONK_API_KEYCLOAK_USER']
 KEYCLOAK_USER_PASSWORD: str = os.environ['SQUONK_API_KEYCLOAK_USER_PASSWORD']
+# Optional
+API_URL_VALIDATION: bool = os.environ.get('SQUONK_API_URL_VALIDATION', 'yes').lower() == 'yes'
+
+
+def fail(msg: str, rv: Optional[DmApiRv] = None) -> NoReturn:
+    """Issues a failure message then sies a sys.exit(1).
+    """
+    err_msg = f'FAILED {msg}'
+    if rv:
+        assert not rv.success
+        err_msg += f' (rv.msg={rv.msg})'
+    print(err_msg)
+    sys.exit(1)
 
 
 def main():
@@ -30,13 +45,15 @@ def main():
     # Configure.
     # Depending on keycloak configuration
     # you may only have 5 minutes before the token expires.
-    DmApi.set_api_url(API_URL)
+    DmApi.set_api_url(API_URL, verify_ssl_cert=API_URL_VALIDATION)
     url, verify = DmApi.get_api_url()
     assert url == API_URL
     first_token = DmApi.get_access_token(
         KEYCLOAK_URL, KEYCLOAK_REALM, KEYCLOAK_CLIENT_ID,
         KEYCLOAK_USER, KEYCLOAK_USER_PASSWORD)
-    assert first_token
+    if not first_token:
+        fail(f'Failed to get token from "{KEYCLOAK_URL}"')
+        sys.exit(1)
     print(f'DM-API connected ({API_URL})')
 
     # Get another token using the existing token.
@@ -49,14 +66,17 @@ def main():
 
     # Basic ping/version
     rv: DmApiRv = DmApi.ping(token)
-    assert rv.success
+    if not rv.success:
+        fail('ping()', rv)
     rv = DmApi.get_version(token)
-    assert rv.success
+    if not rv.success:
+        fail(f'get_version()', rv)
     print(f"DM-API version={rv.msg['version']}")
 
     # Get existing projects
     rv_projects = DmApi.get_available_projects(token)
-    assert rv_projects.success
+    if not rv_projects.success:
+        fail('get_available_projects()', rv)
 
     # If given a project ID find it,
     # otherwise create one using the '1111' codes,
@@ -69,11 +89,13 @@ def main():
                 print(f"Found project (product_id={project['product_id']} size={project['size']})")
                 found = True
                 break
-        assert found
+        if not found:
+            fail(f'Project does not exist ({args.project_id})')
         project_id = args.project_id
     else:
-        # Delete project if it exists
-        new_project_name = 'DmApi Project'
+        # Crete a new project, using the one we're about to
+        # create if it already exists on the server.
+        new_project_name = 'DmApi Test Project'
         project_exists = False
         for project in rv_projects.msg['projects']:
             if project['name'] == new_project_name:
@@ -81,15 +103,21 @@ def main():
                 project_id = project['project_id']
                 project_exists = True
                 break
-        if not project_exists:
-            print(f'Creating new Project ("{new_project_name}")...')
-            rv = DmApi.create_project(token, 'DmApi-generated Project')
-            assert rv.success
+        if project_exists:
+            print(f'Test project ("{new_project_name}") already exists')
+        else:
+            # By not providing a project ID
+            # we create one using a build-in Account Server "Product"
+            # specifically designed for testing.
+            print(f'Creating new Project ("{new_project_name}", "{TEST_PRODUCT_ID})...')
+            rv = DmApi.create_project(token, new_project_name, TEST_PRODUCT_ID)
+            if not rv.success:
+                fail('create_project()', rv)
             project_id = rv.msg['project_id']
             print(f'Created project_id={project_id}')
     assert project_id
 
-    # get a list of project files on the root
+    # Get a list of project files on the root
     print("Listing project files")
     rv = DmApi.list_project_files(token, project_id, '/')
     assert rv.success
@@ -110,16 +138,19 @@ def main():
     rv = DmApi.put_unmanaged_project_files(token, project_id,
                                            local_file,
                                            project_path=project_path)
-    assert rv.success
+    if not rv.success:
+        fail(f'put_unmanaged_project_files({project_id})', rv)
     rv = DmApi.get_unmanaged_project_file(token, project_id,
                                           local_file,
                                           project_path=project_path,
                                           local_file=local_file)
-    assert rv.success
+    if not rv.success:
+        fail(f'get_unmanaged_project_file({project_id})', rv)
     rv = DmApi.delete_unmanaged_project_files(token, project_id,
                                               local_file,
                                               project_path=project_path)
-    assert rv.success
+    if not rv.success:
+        fail(f'delete_unmanaged_project_files({project_id})', rv)
 
     # Run a test job
     print('Starting Job...')
@@ -127,10 +158,12 @@ def main():
     job_name = 'nop'
     job_version = '1.0.0'
     rv = DmApi.get_job_by_name(token, job_collection, job_name, job_version)
-    assert rv.success
+    if not rv.success:
+        fail('get_job_by_name()', rv)
     spec = {'collection': job_collection, 'job': job_name, 'version': job_version}
-    rv = DmApi.start_job_instance(token, project_id, 'DmApi Job', specification=spec)
-    assert rv.success
+    rv = DmApi.start_job_instance(token, project_id, 'DmApi Test Job', specification=spec)
+    if not rv.success:
+        fail('start_job_instance()', rv)
     job_task_id = rv.msg['task_id']
     job_instance_id = rv.msg['instance_id']
     print(f'Started (task_id={job_task_id} instance_id={job_instance_id})')
@@ -155,12 +188,14 @@ def main():
     assert done
     print('Job events...')
     rv = DmApi.get_task(token, job_task_id)
-    assert rv.success
+    if not rv.success:
+        fail(f'get_task({job_task_id})', rv)
     for event in rv.msg['events']:
         print(f"# {event['time']} {event['level']} {event['message']}")
     print('Deleting Job...')
     rv = DmApi.delete_instance(token, job_instance_id)
-    assert rv.success
+    if not rv.success:
+        fail(f'delete_instance({job_instance_id})', rv)
     print('Deleted')
 
     # Finally, if we created a project, delete it.
@@ -168,7 +203,8 @@ def main():
         print('Deleting project I created...')
         print(f'Deleting project_id={project_id}...')
         rv = DmApi.delete_project(token, project_id)
-        assert rv.success
+        if not rv.success:
+            fail(f'delete_project({project_id})', rv)
         print('Deleted')
 
     print('Done')
